# services/http/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  handlers.BlankStruct:
    type: object
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  handlers.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
    type: object
  users.GenerateUserAccessTokenDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  users.RegisterUserDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  users.UserDTO:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:9000
info:
  contact: {}
  description: A simple JWT powered API written in Go
  title: Simple JWT API Go
  version: "1.0"
paths:
  /blacklist-access-token:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.BlankStruct'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - securitydefinitions.apikey: []
      summary: Blacklist access token for user
  /generate-access-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/users.GenerateUserAccessTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/handlers.BlankStruct'
                  - properties:
                      access_token:
                        type: string
                      type:
                        type: string
                      user:
                        $ref: '#/definitions/users.UserDTO'
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Generate access token for a new user
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/users.UserDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - securitydefinitions.apikey: []
      summary: Get authenticated user
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/users.RegisterUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/users.UserDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a new user
produces:
- application/json
swagger: "2.0"